+++ __docs__/components/gallery/tiles/breadcrumbs-tile.tsx
@@ -25,12 +25,14 @@ export default function BreadcrumbsTile(props: CommonTileProps) {
             <View style={styles.centerContent}>
                 <Breadcrumbs>
                     <BreadcrumbsItem>
-                        {/* eslint-disable-next-line jsx-a11y/anchor-is-valid -- TODO: Address a11y error */}
-                        <Link href="">Course</Link>
+                        <Link href="#course" target="_blank">
+                            Course
+                        </Link>
                     </BreadcrumbsItem>
                     <BreadcrumbsItem>
-                        {/* eslint-disable-next-line jsx-a11y/anchor-is-valid -- TODO: Address a11y error */}
-                        <Link href="">Unit</Link>
+                        <Link href="#unit" target="_blank">
+                            Unit
+                        </Link>
                     </BreadcrumbsItem>
                     <BreadcrumbsItem>Lesson</BreadcrumbsItem>
                 </Breadcrumbs>