+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 6.1.2
+
+### Patch Changes
+
+- 5655b9f: Switch to use `focus.outer` semanticColor token
+- 8fc65a9: Migrate Clickable to `semanticColor`
+- 051f0f8: Rework action.disabled tokens to include properties per kind (border, background, foreground).
+- Updated dependencies [ed26d66]
+- Updated dependencies [5655b9f]
+- Updated dependencies [5655b9f]
+- Updated dependencies [8f53293]
+- Updated dependencies [051f0f8]
+- Updated dependencies [e1b78db]
+- Updated dependencies [051f0f8]
+    - @khanacademy/wonder-blocks-core@12.2.0
+    - @khanacademy/wonder-blocks-tokens@5.0.0
+
 ## 6.1.1
 
 ### Patch Changes