+++ packages/wonder-blocks-testing/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-testing
 
+## 15.1.2
+
+### Patch Changes
+
+- Updated dependencies [ed26d66]
+- Updated dependencies [eb47d37]
+    - @khanacademy/wonder-blocks-core@12.2.0
+    - @khanacademy/wonder-blocks-testing-core@2.2.0
+    - @khanacademy/wonder-blocks-data@14.1.2
+
 ## 15.1.1
 
 ### Patch Changes